+++ src/client/sound/s_mix.h
@@ -23,7 +23,7 @@
 #define __S_MIX_H__
 
 void S_LoopSample(const vec3_t org, s_sample_t *sample);
-void S_PlaySample(const vec3_t org, int ent_num, s_sample_t *sample, int atten);
+void S_PlaySample(const vec3_t org, unsigned short ent_num, s_sample_t *sample, int atten);
 void S_StartLocalSample(const char *name);
 
 #ifdef __S_LOCAL_H__
